pages/base_page.py
import allure
from selenium.webdriver.support.wait import WebDriverWait
from selenium.webdriver.support import expected_conditions


class BasePage:

    def __init__(self, driver):
        self.driver = driver

    @allure.step('Получить текущий адрес страницы')
    def get_current_url(self):
        return self.driver.current_url

    @allure.step('Дождаться, пока поменяется страница')
    def wait_url_changes(self, url):
        WebDriverWait(self.driver, 20).until(expected_conditions.url_changes(url))

    @allure.step('Дождаться нужного элемента по локатору')
    def wait_and_find_element(self, locator):
        WebDriverWait(self.driver, 10).until(expected_conditions.visibility_of_element_located(locator))
        return self.driver.find_element(*locator)

    @allure.step('Дождаться, пока нужный элемент по локатору не исчезнет')
    def wait_and_find_element_invisible(self, locator):
        WebDriverWait(self.driver, 10).until(expected_conditions.invisibility_of_element_located(locator))
        return self.driver.find_element(*locator)

    @allure.step('Найти элемент и вытащить текст элемента')
    def find_text(self, locator):
        element = self.driver.find_element(*locator)
        return element.text

    @allure.step('Кликнуть по элементу с нужным локатором')
    def click(self, locator):
        button = self.driver.find_element(*locator)
        self.driver.execute_script("arguments[0].click();", button)
